+++ src/utsuho/__init__.py
@@ -1,5 +1,5 @@
 """ Utsuho is an interconverter between Japanese half-width katakana and full-width katakana.
 """
-from .converters import ConverterConfig, FullToHalfConverter, HalfToFullConverter
+from .converters import WidthConverterConfig, FullToHalfConverter, HalfToFullConverter
 
-__version__ = '1.1.2'
+__version__ = '2.0.0'