+++ src/utsuho/cli.py
@@ -5,14 +5,16 @@ import os.path
 import click
 
 from . import __version__
-from .converters import FullToHalfConverter, HalfToFullConverter
+from .converters import FullToHalfConverter, HalfToFullConverter, HiraganaToKatakanaConverter, KatakanaToHiraganaConverter
 
 
 @click.group(invoke_without_command=True)
 @click.option('--version', is_flag=True, help='Show version.')
 @click.pass_context
 def cli(ctx: click.Context, version: bool):
-    """ Utsuho is an interconverter between Japanese half-width katakana and full-width katakana.\f
+    """ Utsuho is a Python module that facilitates bidirectional conversion
+    between half-width katakana and full-width katakana in Japanese,
+    as well as between hiragana and katakana.\f
 
     Parameters
     ----------
@@ -31,7 +33,10 @@ def cli(ctx: click.Context, version: bool):
 
 
 @cli.command()
-@click.option('--file', '-f', 'file_', is_flag=True, help='Use TEXT as file path.')
+@click.option(
+    '--file', '-f', 'file_', is_flag=True,
+    help='Whether to use TEXT as a file path or not.'
+)
 @click.argument('text')
 def full_to_half(file_: bool, text: str):
     """ Convert full-width katakana to half-width katakana.\f
@@ -55,7 +60,10 @@ def full_to_half(file_: bool, text: str):
 
 
 @cli.command()
-@click.option('--file', '-f', 'file_', is_flag=True, help='Use TEXT as file path.')
+@click.option(
+    '--file', '-f', 'file_', is_flag=True,
+    help='Whether to use TEXT as a file path or not.'
+)
 @click.argument('text')
 def half_to_full(file_: bool, text: str):
     """ Convert half-width katakana to full-width katakana.\f
@@ -74,5 +82,59 @@ def half_to_full(file_: bool, text: str):
         s = text
 
     cnv = HalfToFullConverter()
+    converted = cnv.convert(s)
+    click.echo(converted)
+
+
+@cli.command()
+@click.option(
+    '--file', '-f', 'file_', is_flag=True,
+    help='Whether to use TEXT as a file path or not.'
+)
+@click.argument('text')
+def hiragana_to_katakana(file_: bool, text: str):
+    """ Convert hiragana to katakana.\f
+
+    Parameters
+    ----------
+    file_: bool
+        Whether to treat TEXT as a file path.
+    text: str
+        String containing characters to convert to katakana or the path of file containing them.
+    """
+    if file_:
+        with open(os.path.abspath(text), 'r') as fp:
+            s = fp.read()
+    else:
+        s = text
+
+    cnv = HiraganaToKatakanaConverter()
+    converted = cnv.convert(s)
+    click.echo(converted)
+
+
+@cli.command()
+@click.option(
+    '--file', '-f', 'file_', is_flag=True,
+    help='Whether to use TEXT as a file path or not.'
+)
+@click.argument('text')
+def katakana_to_hiragana(file_: bool, text: str):
+    """ Convert katakana to hiragana.\f
+
+    Parameters
+    ----------
+    file_: bool
+        Whether to treat TEXT as a file path.
+    text: str
+        String containing characters to convert to hiragana or the path of file containing them.
+    """
+    if file_:
+        with open(os.path.abspath(text), 'r') as fp:
+            s = fp.read()
+    else:
+        s = text
+
+    cnv = KatakanaToHiraganaConverter()
     converted = cnv.convert(s)
     click.echo(converted)